# sdd/contracts/openapi.yaml
openapi: 3.0.3
info:
  title: 任务管理系统 API
  description: |
    任务管理系统后端 API 文档
    
    核心功能：
    - Excel 导入任务数据
    - 模板管理
    - Rundown 发布清单管理
    - 任务执行控制
    - Jenkins/Ansible 自动化集成
    - 定时任务调度
    - 批量执行
  version: 1.0.0
  contact:
    name: API Support
servers:
  - url: http://localhost:8000/api/v1
    description: 本地开发服务器
  - url: http://{host}:{port}/api/v1
    description: 生产服务器
    variables:
      host:
        default: localhost
      port:
        default: "8000"
...
paths:
...
  /templates:
    get:
      tags:
        - Templates
      summary: 获取模板列表
      operationId: listTemplates
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: page_size
          in: query
          schema:
            type: integer
            default: 20
      responses:
        '200':
          description: 模板列表
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedTemplates'

    post:
      tags:
        - Templates
      summary: 创建新模板
      operationId: createTemplate
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateTemplateRequest'
      responses:
        '201':
          description: 创建成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Template'
  /templates/{template_id}:
    get:
      tags:
        - Templates
      summary: 获取模板详情
      operationId: getTemplate
      parameters:
        - name: template_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: 模板详情
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TemplateDetail'

    put:
      tags:
        - Templates
      summary: 更新模板
      operationId: updateTemplate
      parameters:
        - name: template_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateTemplateRequest'
      responses:
        '200':
          description: 更新成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Template'

    delete:
      tags:
        - Templates
      summary: 删除模板
      operationId: deleteTemplate
      parameters:
        - name: template_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '204':
          description: 删除成功
  /templates/{template_id}/clone:
    post:
      tags:
        - Templates
      summary: 克隆模板
      operationId: cloneTemplate
      parameters:
        - name: template_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
              properties:
                new_name:
                  type: string
      responses:
        '201':
          description: 克隆成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Template'
  /templates/{template_id}/generate-rundown:
    post:
      tags:
        - Templates
      summary: 从模板生成 Rundown
      operationId: generateRundown
      parameters:
        - name: template_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/GenerateRundownRequest'
      responses:
        '201':
          description: 生成成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Rundown'

  # ==================== Rundowns ====================
...
components:
  securitySchemes:
    CookieAuth:
      type: apiKey
      in: cookie
      name: session_id
  schemas:
...
    PaginatedTemplates:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/Template'
        total:
          type: integer
        page:
          type: integer
        page_size:
          type: integer
    Template:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        description:
          type: string
        created_by:
          type: string
          format: uuid
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    TemplateDetail:
      allOf:
        - $ref: '#/components/schemas/Template'
        - type: object
          properties:
            tasks:
              type: array
              items:
                $ref: '#/components/schemas/TemplateTask'
    TemplateTask:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        task_type:
          type: string
          enum: [jenkins, ansible]
        config:
          type: object
        order_index:
          type: integer
    CreateTemplateRequest:
      type: object
      required:
        - name
        - tasks
      properties:
        name:
          type: string
        description:
          type: string
        tasks:
          type: array
          items:
            $ref: '#/components/schemas/TemplateTask'
    UpdateTemplateRequest:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
        tasks:
          type: array
          items:
            $ref: '#/components/schemas/TemplateTask'
    GenerateRundownRequest:
      type: object
      required:
        - name
      properties:
        name:
          type: string
        description:
          type: string

    # Rundowns
...
    Rundown:
      type: object
      properties:
        id:
          type: string
          format: uuid
        rundown_code:
          type: string
        name:
          type: string
        description:
          type: string
        template_id:
          type: string
          format: uuid
        status:
          type: string
          enum: [pending, running, partial_success, success, aborted, failed]
        created_by:
          type: string
          format: uuid
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
        started_at:
          type: string
          format: date-time
        finished_at:
          type: string
          format: date-time
...